extract flux events into separate file
